feat: add cell_type_classifier function
This commit adds the `cell_type_classifier` function to the main file. This function takes an expression set of a mixture of cell expression data and a pretrained model and returns a DataFrame with the predicted cell types. The model should be a GraphNeuralNetworks model. The DataFrame should have the columns "cell_id" and "cell_type", with the "cell_id" column having the same order as the input expression set and the "cell_type" column containing the predicted cell type for each cell.